refactor: Remove value from bytecode project
