+++ bytecode/CMakeLists.txt
@@ -9,5 +9,3 @@ add_library(
 target_include_directories(
     fluir.code INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
-
-add_subdirectory(test)
